Refactor manage-tasks
