+++ frontend/src/App.tsx
@@ -1,6 +1,6 @@
 import Modal from "./components/common/modal/Modal";
-import SupervisorTaskManagement from "./components/supervisor-task-management/SupervisorTaskManagement";
-import { TaskProvider } from "./data/contexts/TaskContext";
+import SupervisorTaskManagement from "./components/manage-tasks/TaskManager";
+import { TaskProvider } from "./contexts/TaskContext";
 
 const App = () => {
   return (
