+++ frontend/src/components/manage-tasks/filter/Filter.tsx
@@ -1,0 +1,9 @@
+const SearchFilter = () => {
+  return (
+    <h1 className="text-3xl text-slate-950 font-black uppercase">
+      Task Manager
+    </h1>
+  );
+};
+
+export default SearchFilter;
