--- frontend/src/components/supervisor-task-management/SearchFilter.tsx
@@ -1,14 +1,0 @@
-const SearchFilter = () => {
-  return (
-    <section
-      className="bg-slate-500 flex flex-row justify-center items-center"
-      style={{ height: "20%" }}
-    >
-      <h1 className="text-3xl text-slate-950 font-black uppercase">
-        Task Manager
-      </h1>
-    </section>
-  );
-};
-
-export default SearchFilter;
