--- frontend/src/components/supervisor-task-management/SupervisorTaskManagement.tsx
@@ -1,15 +1,0 @@
-import SearchFilter from "./SearchFilter";
-import TaskList from "./TaskList";
-import Navigation from "./Navigation";
-
-const SupervisorTaskManagement = () => {
-  return (
-    <main className="h-screen">
-      <SearchFilter />
-      <TaskList />
-      <Navigation />
-    </main>
-  );
-};
-
-export default SupervisorTaskManagement;
