--- frontend/src/data/README.md
@@ -1,7 +1,0 @@
-# Data section
-
-The data section manages the interaction with api and the storage of data.
-It contains contexts, models and api helper methods.
-Services contains request methods for communicating with the api.
-Models are interfaces used in communicating wit the api.
-Contexts manage the data of the application and communicate with the api.
